refactor: move discovery to lib/discovery/
