+++ src/lib/discovery/index.ts
@@ -1,0 +1,23 @@
+export {
+  discover,
+  filterByType,
+  filterByFormat,
+  type DiscoverOptions,
+  type DiscoveryWarning,
+  type DiscoveryResult,
+} from './context-discovery.ts';
+
+export {
+  shouldInstallInternalSkills,
+  parseSkillMd,
+  discoverSkills,
+  getSkillDisplayName,
+  filterSkills,
+  type DiscoverSkillsOptions,
+} from './skill-discovery.ts';
+
+export {
+  discoverRemoteContext,
+  type RemoteContextItem,
+  type RemoteContextSummary,
+} from './find-discovery.ts';
