+++ src/lib/discovery/instruction-discovery.test.ts
