+++ tests/skill-matching.test.ts
@@ -9,7 +9,7 @@ import { describe, it, expect, beforeEach, afterEach } from 'vitest';
 import { mkdirSync, rmSync, writeFileSync } from 'fs';
 import { join } from 'path';
 import { tmpdir } from 'os';
-import { filterSkills, parseSkillMd } from '../src/skill-discovery.ts';
+import { filterSkills, parseSkillMd } from '../src/lib/discovery/index.ts';
 import type { Skill } from '../src/types.ts';
 
 // Mock skill factory
